-change Integer to Long

+++ src/main/java/org/drugref/ca/dpd/CdRoute.java
@@ -51,12 +51,12 @@ public class CdRoute implements Serializable {
     private String routeOfAdministration;
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     public CdRoute() {
     }
 
-    public CdRoute(Integer id) {
+    public CdRoute(Long id) {
         this.id = id;
     }
 
@@ -84,11 +84,11 @@ public class CdRoute implements Serializable {
         this.routeOfAdministration = routeOfAdministration;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
